requirements.txt
Flask
pycups
PyPDF2
